fix(rtx): wavefront panel-review fixes (QMC frameID, 64-bit waves, checkerboard spp)
Five-reviewer panel on the wavefront slices. Fixes the confirmed issues:

- HIGH: the camera QMC (Halton) index used only the per-frame slot
  ordinal (0..spp-1), omitting frameID. Every progressive accumulation
  frame re-cast identical primary rays, so AA/DoF never refined and
  sampleLimit was a no-op. Feed makePrimaryRay frameID + slot.sampleIdx
  (the sample's accumulated ordinal), and key the RNG seed on the same.
- HIGH: totalSamples = numPixels*spp and the waveBase loop counter were
  uint32; a large frame x spp overflowed to a black frame, and a
  waveBase that crossed 2^32 mid-loop wrapped to 0 -> infinite host
  loop. Widened the sample arithmetic (host loop + regenerate kernel) to
  64-bit.
- LOW: the base commitParameters clamps spp to 1 when checkerboarding is
  requested; Wavefront disabled checkerboarding only afterward, so
  checkerboarding=true + pixelSamples=N silently rendered at 1 spp. Now
  restore the requested pixelSamples after disabling checkerboarding.

Nits: corrected the poolScreenSample 'mirrors createScreenSample
exactly' comment; annotated the JSON checkerboarding param as ignored;
bumped the README driver requirement to 570+ (OptiX 9.0 default).

The QMC fix is inspection-verified (unanimous panel) — no automated
regression: this API test harness resets accumulation every render()
(numSamples stays 1 for wavefront AND quality; manual accumulationVersion
did not change it), so cross-frame refinement can't be exercised here.
85/85 tests pass.

+++ devices/rtx/device/renderer/WavefrontLaunch.cu
@@ -38,23 +38,23 @@ namespace {
 constexpr int kThreadsPerBlock = 256;
 
 __global__ void wavefrontRegenerateKernel(WavefrontPathSlot *slots,
-    uint32_t waveBase,
+    uint64_t waveBase,
     uint32_t numPixels,
-    uint32_t totalSamples,
+    uint64_t totalSamples,
     uint32_t liveSlots)
 {
   const uint32_t i = blockIdx.x * blockDim.x + threadIdx.x;
   if (i >= liveSlots)
     return;
 
-  const uint32_t sampleId = waveBase + i;
+  const uint64_t sampleId = waveBase + i;
   if (sampleId >= totalSamples) {
     slots[i].alive = 0;
     return;
   }
 
-  slots[i].pixel = sampleId % numPixels;
-  slots[i].sampleIdx = sampleId / numPixels;
+  slots[i].pixel = uint32_t(sampleId % numPixels);
+  slots[i].sampleIdx = uint32_t(sampleId / numPixels);
   slots[i].alive = 1;
 }
 
@@ -62,9 +62,9 @@ __global__ void wavefrontRegenerateKernel(WavefrontPathSlot *slots,
 
 void wavefrontRegenerate(cudaStream_t stream,
     WavefrontPathSlot *slots,
-    uint32_t waveBase,
+    uint64_t waveBase,
     uint32_t numPixels,
-    uint32_t totalSamples,
+    uint64_t totalSamples,
     uint32_t liveSlots)
 {
   if (liveSlots == 0)
